renamed getLayerNumberProperties to getLayerProperties and to be able to get string columns for label
