+++ src/components/controls/VectorLineAdvanced.svelte
@@ -19,7 +19,7 @@
   } from '$lib/constants'
   import {
     getIntervalList,
-    getLayerNumberProperties,
+    getLayerProperties,
     getLayerStyle,
     getLineWidth,
     getSampleFromInterval,
@@ -98,7 +98,7 @@
   }
 
   const getPropertySelectValue = () => {
-    const vectorLayerMeta = getLayerNumberProperties($map, layer)
+    const vectorLayerMeta = getLayerProperties($map, layer)
     const selectOptions = Object.keys(vectorLayerMeta.fields)
 
     propertySelectValue = selectOptions[0]
