+++ src/components/controls/VectorPolygonAdvanced.svelte
@@ -26,7 +26,7 @@
   import { map } from '$stores'
   import {
     getIntervalList,
-    getLayerNumberProperties,
+    getLayerProperties,
     getLayerStyle,
     getSampleFromInterval,
     remapInputValue,
@@ -70,7 +70,7 @@
   })
 
   const getPropertySelectValue = () => {
-    const vectorLayerMeta = getLayerNumberProperties($map, layer)
+    const vectorLayerMeta = getLayerProperties($map, layer)
     const selectOptions = Object.keys(vectorLayerMeta.fields)
 
     propertySelectValue = selectOptions[0]
