+++ src/lib/helper/index.ts
@@ -8,7 +8,7 @@ export * from './generateMetadataJson'
 export * from './getActiveBandIndex'
 export * from './getBase64EncodedUrl'
 export * from './getIntervalList'
-export * from './getLayerNumberProperties'
+export * from './getLayerProperties'
 export * from './getLayerStyle'
 export * from './getLayerUrl'
 export * from './getLineWidth'
